requirements.txt
phi
groq
yfinance
duckduckgo-search
python-dotenv
requests
pandas
